requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.7.4
anyio==3.3.3
apache-airflow==2.2.1
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-redis==2.0.1
apache-airflow-providers-sftp==4.1.0
apache-airflow-providers-slack==4.1.0
apache-airflow-providers-sqlite==2.0.1
apache-airflow-providers-ssh==3.2.0
apispec==3.3.2
apturl==0.5.2
argcomplete==1.12.3
arrow==1.2.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==20.1.0
Automat==20.2.0
aws-secretsmanager-caching==1.1.1.5
awscli==1.27.159
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==23.7.0
blinker==1.4
boto3==1.26.76
botocore==1.29.76
Brlapi==0.8.3
build==0.10.0
CacheControl==0.12.14
cachelib==0.10.2
cachetools==5.3.1
cachy==0.3.0
cattrs==1.5.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
cleo==2.0.1
click==8.1.6
clickclick==20.10.2
clikit==0.6.2
cloudpickle==2.2.1
colorama==0.4.4
colorlog==4.8.0
comm==0.1.3
command-not-found==0.3
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==15.1.0
contourpy==1.1.0
crashtest==0.4.1
cron-descriptor==1.4.0
croniter==1.0.15
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
docutils==0.16
dulwich==0.21.5
duplicity==0.8.21
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.12.0
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Limiter==3.3.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.42.0
frozenlist==1.3.3
future==0.18.2
git-remote-codecommit==1.16
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-play-scraper==1.2.0
googleapis-common-protos==1.59.1
graphviz==0.17
greenlet==1.1.2
gspread==5.10.0
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httpcore==0.13.7
httplib2==0.20.2
httpx==0.19.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==21.3.0
inflection==0.5.1
installer==0.7.0
ipykernel==6.25.0
ipython==8.14.0
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
itunes-app-scraper-dmi==0.9.5
jaraco.classes==3.2.3
jedi==0.19.0
jeepney==0.7.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.3.1
jsonschema==3.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kaleido==0.2.1
keyring==23.13.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.4.3
ldap3==2.8.1
limits==3.5.0
linkify-it-py==2.0.2
llvmlite==0.40.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.5
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
missingno==0.5.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==1.4.6
nbformat==5.9.2
nest-asyncio==1.5.7
netaddr==0.8.0
netifaces==0.11.0
numba==0.57.1
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==1.3.4
paramiko==2.9.3
parso==0.8.3
pastel==0.2.1
pathspec==0.9.0
pbr==5.8.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.9.6
platformdirs==3.5.1
plotly==5.16.0
pluggy==1.2.0
ply==3.11
poetry==1.5.0
poetry-core==1.6.0
poetry-plugin-export==1.4.0
prison==0.2.1
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.9
Pygments==2.10.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylev==1.2.0
pymacaroons==0.13.0
pymongo==4.4.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyRFC3339==1.1
pyrsistent==0.18.0
pysftp==0.2.9
pyspark==3.4.0
python-apt==2.4.0+ubuntu1
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
rapidfuzz==2.15.1
redis==3.5.3
reportlab==3.6.8
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==10.12.0
rich_argparse==1.1.1
roman==3.3
rsa==4.7.2
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.2.2
shap==0.41.0
shellingham==1.5.0.post1
six==1.16.0
slack-sdk==3.21.3
slicer==0.0.7
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
sqlalchemy-migrate==0.13.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.2
sshtunnel==0.4.0
stack-data==0.6.2
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.9
Tempita==0.5.2
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.5.24
Twisted==22.1.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
virtualenv==20.23.0
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.15.0
WTForms==2.3.3
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
